requirements.txt
# Core dependencies
google-api-python-client==2.108.0
google-auth==2.25.2
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
openai==1.6.1
requests==2.31.0
pyyaml==6.0.1

# Video processing
ffmpeg-python==0.2.0

# Data handling
pandas==2.1.4
numpy==1.26.3

# Utilities
python-dotenv==1.0.0
tenacity==8.2.3
